add U2EInformation component and its test
