+++ app/src/pages/AppSettings/__test__/AdvancedSettings.test.tsx
@@ -23,12 +23,11 @@ import {
 import * as CustomLabware from '../../../redux/custom-labware'
 import * as Config from '../../../redux/config'
 import * as ProtocolAnalysis from '../../../redux/protocol-analysis'
-import * as SystemInfo from '../../../redux/system-info'
-import * as Fixtures from '../../../redux/system-info/__fixtures__'
 import {
   EnableDevTools,
   OT2AdvancedSettings,
   PreventRobotCaching,
+  U2EInformation,
 } from '../../../organisms/AdvancedSettings'
 
 import { AdvancedSettings } from '../AdvancedSettings'
@@ -74,14 +73,6 @@ const mockGetIsLabwareOffsetCodeSnippetsOn = Config.getIsLabwareOffsetCodeSnippe
   typeof Config.getIsLabwareOffsetCodeSnippetsOn
 >
 
-const mockGetU2EAdapterDevice = SystemInfo.getU2EAdapterDevice as jest.MockedFunction<
-  typeof SystemInfo.getU2EAdapterDevice
->
-
-const mockGetU2EWindowsDriverStatus = SystemInfo.getU2EWindowsDriverStatus as jest.MockedFunction<
-  typeof SystemInfo.getU2EWindowsDriverStatus
->
-
 const mockGetIsHeaterShakerAttached = Config.getIsHeaterShakerAttached as jest.MockedFunction<
   typeof Config.getIsHeaterShakerAttached
 >
@@ -107,6 +98,9 @@ const mockOT2AdvancedSettings = OT2AdvancedSettings as jest.MockedFunction<
 const mockEnableDevTools = EnableDevTools as jest.MockedFunction<
   typeof EnableDevTools
 >
+const mockU2EInformation = U2EInformation as jest.MockedFunction<
+  typeof U2EInformation
+>
 
 let mockTrackEvent: jest.Mock
 const mockConfirm = jest.fn()
@@ -125,10 +119,8 @@ describe('AdvancedSettings', () => {
       { label: 'Beta', value: 'beta' },
       { label: 'Alpha', value: 'alpha' },
     ])
-    mockGetU2EAdapterDevice.mockReturnValue(Fixtures.mockWindowsRealtekDevice)
     mockGetUnreachableRobots.mockReturnValue([mockUnreachableRobot])
     mockGetReachableRobots.mockReturnValue([mockReachableRobot])
-    mockGetU2EWindowsDriverStatus.mockReturnValue(SystemInfo.OUTDATED)
     mockUseConditionalConfirm.mockReturnValue({
       confirm: mockConfirm,
       showConfirmation: true,
@@ -137,6 +129,7 @@ describe('AdvancedSettings', () => {
     mockPreventRobotCaching.mockReturnValue(<div>mock PreventRobotCaching</div>)
     mockOT2AdvancedSettings.mockReturnValue(<div>mock OT2AdvancedSettings</div>)
     mockEnableDevTools.mockReturnValue(<div>mock EnableDevTools</div>)
+    mockU2EInformation.mockReturnValue(<div>mock U2EInformation</div>)
   })
 
   afterEach(() => {
@@ -149,7 +142,6 @@ describe('AdvancedSettings', () => {
     getByText('Update Channel')
     getByText('Additional Custom Labware Source Folder')
     getByText('Clear Unavailable Robots')
-    getByText('USB-to-Ethernet Adapter Information')
   })
 
   it('renders the update channel combobox and section', () => {
@@ -190,68 +182,9 @@ describe('AdvancedSettings', () => {
     screen.getByText('mock PreventRobotCaching')
   })
 
-  it('render the usb-to-ethernet adapter information', () => {
-    const [{ getByText }] = render()
-    getByText('USB-to-Ethernet Adapter Information')
-    getByText(
-      'Some OT-2s have an internal USB-to-Ethernet adapter. If your OT-2 uses this adapter, it will be added to your computer’s device list when you make a wired connection. If you have a Realtek adapter, it is essential that the driver is up to date.'
-    )
-    getByText('Description')
-    getByText('Manufacturer')
-    getByText('Driver Version')
-  })
-
-  it('renders the test data of the usb-to-ethernet adapter information with mac', () => {
-    mockGetU2EAdapterDevice.mockReturnValue({
-      ...Fixtures.mockRealtekDevice,
-    })
-    mockGetU2EWindowsDriverStatus.mockReturnValue(SystemInfo.NOT_APPLICABLE)
-    const [{ getByText, queryByText }] = render()
-    getByText('USB 10/100 LAN')
-    getByText('Realtek')
-    getByText('Unknown')
-    expect(
-      queryByText(
-        'An update is available for Realtek USB-to-Ethernet adapter driver'
-      )
-    ).not.toBeInTheDocument()
-    expect(queryByText('go to Realtek.com')).not.toBeInTheDocument()
-  })
-
-  it('renders the test data of the outdated usb-to-ethernet adapter information with windows', () => {
-    const [{ getByText }] = render()
-    getByText('Realtek USB FE Family Controller')
-    getByText('Realtek')
-    getByText('1.2.3')
-    getByText(
-      'An update is available for Realtek USB-to-Ethernet adapter driver'
-    )
-    const targetLink = 'https://www.realtek.com/en/'
-    const link = getByText('go to Realtek.com')
-    expect(link.closest('a')).toHaveAttribute('href', targetLink)
-  })
-
-  it('renders the test data of the updated usb-to-ethernet adapter information with windows', () => {
-    mockGetU2EWindowsDriverStatus.mockReturnValue(SystemInfo.UP_TO_DATE)
-    const [{ getByText, queryByText }] = render()
-    getByText('Realtek USB FE Family Controller')
-    getByText('Realtek')
-    getByText('1.2.3')
-    expect(
-      queryByText(
-        'An update is available for Realtek USB-to-Ethernet adapter driver'
-      )
-    ).not.toBeInTheDocument()
-    expect(queryByText('go to Realtek.com')).not.toBeInTheDocument()
-  })
-
-  it('renders the not connected message and not display item titles when USB-to-Ethernet is not connected', () => {
-    mockGetU2EAdapterDevice.mockReturnValue(null)
-    const [{ getByText, queryByText }] = render()
-    expect(queryByText('Description')).not.toBeInTheDocument()
-    expect(queryByText('Manufacturer')).not.toBeInTheDocument()
-    expect(queryByText('Driver Version')).not.toBeInTheDocument()
-    getByText('No USB-to-Ethernet adapter connected')
+  it('should render mock U2EInformation', () => {
+    render()
+    expect(screen.getByText('mock U2EInformation'))
   })
 
   it('renders the display show link to get labware offset data section', () => {
